requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
tensorflow
jinja2
python-multipart
psycopg2-binary